move test to shim
